used a conversion function "float64()". but is it needed?

+++ flow/8.go
@@ -3,7 +3,7 @@ package main
 import "fmt"
 
 func Sqrt(x float64) float64 {
-	r := x
+	r := float64(x)
 	for i := 1; i <= 10; i++ {
 		r = r - (r*r-x)/(2*r)
